feat: set up dependabot
Add a dependabot script for building and testing the project on a PR.

Add dependabot scripts for codeql and dependency scanning.

Load the android cache name from a test argument.

Ignore expect/actual class warnings in project.

Fix two minor warnings for unnecessary casts.

+++ src/androidMain/kotlin/software/momento/kotlin/sdk/internal/InternalTopicClient.android.kt
@@ -159,7 +159,7 @@ internal actual class InternalTopicClient actual constructor(
             this.cacheName = cacheName
             this.topic = topicName
             if (lastSequenceNumber != null) {
-                this.resumeAtTopicSequenceNumber = lastSequenceNumber as Long
+                this.resumeAtTopicSequenceNumber = lastSequenceNumber
             }
         }.build()
     }
